ji_find {emo}	R Documentation

List all emoji with a given keyword

Description

Note that this is unlikely to print correctly on your R console, but it will work in (e.g.) the RStudio viewer.

Usage

ji_find(keyword)

Arguments

keyword
Emoji keyword

Examples

emo::ji_find("happy")

[Package emo version 0.0.0.9000 Index]